pyproject.toml
[project]
name = "pybrag"
dynamic = ["version"]
description = "Basic RAG"
readme = "README.md"
authors = [
    { name = "Arthur Lui", email = "alui@lanl.gov" }
]
requires-python = ">=3.10"
dependencies = [
    # Webapp.
    "gradio~=5.33.0",
    "fastapi>=0.115.7,<0.116",
    # langchain.
    "langchain>=0.3.15",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.15",
    "langchain-core>=0.3.31",
    "langchain-litellm>=0.2.2",
    "langchain-ollama>=0.3.4",
    "langgraph>=0.5.3",
    "docx2txt>=0.9",
    "litellm==1.74.3",
    "openai>=1.60.1,<2.0",
    "pydantic>=2.10.6,<3.0",
    "pymupdf4llm>=0.0.17,<0.1",
    "pypdf>=5.1.0,<6.0",
    "tqdm>=4.67.1,<5.0",
]
license = "MIT"
license-files = [ "LICENSE" ]
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[project.scripts]
brag = "brag:main"

[build-system]
requires = ["setuptools>=68.1", "versioningit"]
build-backend = "setuptools.build_meta"

[tool.versioningit.vcs]
method = "git"
default-tag = "0.0.1"

[tool.ruff]
line-length = 80

[tool.ruff.lint]
ignore = ["D100", "F401"]
extend-select = ["I", "W505"] # "D"
extend-unsafe-fixes = ["F401"]
pydocstyle.convention = "numpy"
pycodestyle.max-doc-length = 80

# Ignore test file documentation linting.
[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = ["D"]

[dependency-groups]
dev = [
    "lark>=1.2.2",
    "llama-index>=0.12.15",
    "llama-index-embeddings-huggingface>=0.5.1",
    "llama-index-llms-ollama>=0.5.0",
    "notebook>=7.3.2",
]
